Made a no-op windows implementation of the linkcache package
