--- pkg/linkcache/cache_test.go
@@ -1,1 +1,0 @@
-package linkcache
