+++ pkg/linkcache/devices_linux.go
@@ -15,18 +15,6 @@ import (
 
 const byIdDir = "/dev/disk/by-id"
 
-type deviceMapping struct {
-	symlink  string
-	realPath string
-}
-
-type DeviceCache struct {
-	volumes map[string]deviceMapping
-	period  time.Duration
-	// dir is the directory to look for device symlinks
-	dir string
-}
-
 func NewDeviceCacheForNode(ctx context.Context, period time.Duration, nodeName string) (*DeviceCache, error) {
 	node, err := k8sclient.GetNodeWithRetry(ctx, nodeName)
 	if err != nil {
